+++ src/binary_tree/tree_builder/single_threaded.rs
